+++ apps/public-docsite-v9-headless/.storybook/preview.js
@@ -8,7 +8,7 @@ export const parameters = {
   options: {
     storySort: {
       method: 'alphabetical',
-      order: ['Introduction', 'Headless Components'],
+      order: ['Overview', ['Introduction', 'Getting Started', 'Accessibility'], 'Guides', 'Components', 'Concepts'],
     },
   },
   reactStorybookAddon: {