+++ packages/react-components/react-headless-components-preview/stories/src/Badge/index.stories.tsx
@@ -4,7 +4,7 @@ import descriptionMd from './BadgeDescription.md';
 export { Default } from './BadgeDefault.stories';
 
 export default {
-  title: 'Headless Components/Badge',
+  title: 'Components/Badge',
   component: Badge,
   parameters: {
     docs: {