+++ packages/web-components/src/compound-button/compound-button.template.ts
@@ -4,12 +4,16 @@ import type { CompoundButton } from './compound-button.js';
 import type { CompoundButtonOptions } from './compound-button.options.js';
 
 /**
- * The template for the Compound Button component.
+ * Generates a template for the Button component.
+ *
  * @public
  */
 export function buttonTemplate<T extends CompoundButton>(options: CompoundButtonOptions = {}): ElementViewTemplate<T> {
   return html<T>`
-    <template ?disabled="${x => x.disabled}" tabindex="${x => (x.disabled ? null : x.tabIndex ?? 0)}">
+    <template
+      @click="${(x, c) => x.clickHandler(c.event as MouseEvent)}"
+      @keypress="${(x, c) => x.keypressHandler(c.event as KeyboardEvent)}"
+    >
       ${startSlotTemplate(options)}
       <span class="content" part="content">
         <slot ${slotted('defaultSlottedContent')}></slot>
@@ -22,6 +26,7 @@ export function buttonTemplate<T extends CompoundButton>(options: CompoundButton
 
 /**
  * The template for the Button component.
+ *
  * @public
  */
 export const template: ElementViewTemplate<CompoundButton> = buttonTemplate();