+++ packages/web-components/src/drawer/drawer.ts
@@ -1,4 +1,4 @@
-import { attr, FASTElement, observable, Updates } from '@microsoft/fast-element';
+import { attr, FASTElement, observable, Updates, volatile } from '@microsoft/fast-element';
 import { DrawerPosition, DrawerSize, DrawerType } from './drawer.options.js';
 
 /**
@@ -33,8 +33,6 @@ import { DrawerPosition, DrawerSize, DrawerType } from './drawer.options.js';
  * @tag fluent-drawer
  */
 export class Drawer extends FASTElement {
-  protected roleAttrObserver!: MutationObserver;
-
   /**
    * Determines whether the drawer should be displayed as modal or non-modal
    * When rendered as a modal, an overlay is applied over the rest of the view.
@@ -42,20 +40,7 @@ export class Drawer extends FASTElement {
    * @public
    */
   @attr
-  public type: DrawerType = DrawerType.modal;
-  protected typeChanged() {
-    if (!this.dialog) {
-      return;
-    }
-
-    this.updateDialogRole();
-
-    if (this.type === DrawerType.modal) {
-      this.dialog.setAttribute('aria-modal', 'true');
-    } else {
-      this.dialog.removeAttribute('aria-modal');
-    }
-  }
+  public type!: DrawerType;
 
   /**
    * The ID of the element that labels the drawer.
@@ -82,6 +67,9 @@ export class Drawer extends FASTElement {
   @attr
   public position: DrawerPosition = DrawerPosition.start;
 
+  @observable
+  public role!: string | null;
+
   /**
    * @public
    * @defaultValue medium
@@ -98,17 +86,79 @@ export class Drawer extends FASTElement {
   @observable
   public dialog!: HTMLDialogElement;
 
-  /** @internal */
-  connectedCallback() {
-    super.connectedCallback();
-    this.typeChanged();
-    this.observeRoleAttr();
+  /**
+   * The `aria-describedby` attribute value for the dialog, which is determined by the `ariaDescribedby` property. This
+   * is used to ensure that the dialog's accessible description is properly announced by assistive technologies.
+   *
+   * @internal
+   */
+  @volatile
+  public get dialogDescribedby(): string | undefined {
+    if (this.dialog) {
+      return this.ariaDescribedby;
+    }
   }
 
-  /** @internal */
-  disconnectedCallback() {
-    super.disconnectedCallback();
-    this.roleAttrObserver.disconnect();
+  /**
+   * The `aria-label` attribute value for the dialog, which is determined by the `ariaLabel` property. This is used to
+   * ensure that the dialog's accessible name is properly announced by assistive technologies.
+   *
+   * @internal
+   */
+  @volatile
+  public get dialogLabel(): string | null | undefined {
+    if (this.dialog) {
+      return this.ariaLabel;
+    }
+  }
+
+  /**
+   * The `aria-labelledby` attribute value for the dialog, which is determined by the `ariaLabelledby` property. This is
+   * used to ensure that the dialog's accessible name is properly announced by assistive technologies.
+   *
+   * @internal
+   */
+  @volatile
+  public get dialogLabelledby(): string | undefined {
+    if (this.dialog) {
+      return this.ariaLabelledby;
+    }
+  }
+
+  /**
+   * The modal state of the dialog, which is determined by the `type` property. If the dialog is not a non-modal dialog,
+   * the modal state will be true, otherwise it will be undefined.
+   *
+   * @internal
+   */
+  @volatile
+  public get dialogModal(): boolean | undefined {
+    if (this.dialog && this.type === DrawerType.modal) {
+      return true;
+    }
+  }
+
+  /**
+   * The role of the dialog, which is determined by the `type` property. If the dialog is an alert dialog, the role will
+   * be 'alertdialog', otherwise it will be undefined.
+   *
+   * @internal
+   */
+  @volatile
+  public get dialogRole(): string | null {
+    if (this.dialog && this.type === DrawerType.modal) {
+      return 'dialog';
+    }
+
+    return this.role;
+  }
+
+  connectedCallback() {
+    super.connectedCallback();
+
+    Updates.enqueue(() => {
+      this.type = this.type ?? DrawerType.modal;
+    });
   }
 
   /**
@@ -185,26 +235,5 @@ export class Drawer extends FASTElement {
    */
   public cancelHandler() {
     this.hide();
-  }
-
-  protected observeRoleAttr() {
-    if (this.roleAttrObserver) {
-      return;
-    }
-
-    this.roleAttrObserver = new MutationObserver(() => {
-      this.updateDialogRole();
-    });
-    this.roleAttrObserver.observe(this, {
-      attributes: true,
-      attributeFilter: ['role'],
-    });
-  }
-
-  protected updateDialogRole() {
-    if (!this.dialog) {
-      return;
-    }
-    this.dialog.role = this.type === DrawerType.modal ? 'dialog' : this.role;
   }
 }